refactor useMediaTrack hooks into separate files
