+++ packages/react/src/hooks/index.ts
@@ -1,8 +1,9 @@
 export * from './participant-hooks';
 export * from './token-hooks';
-export * from './track-hooks';
 export { useDataChannel } from './useDataChannel';
 export { useTracks } from './useTracks';
 export { useGridLayout } from './useGridLayout';
 export { useVisualStableUpdate } from './useVisualStableUpdate';
 export { usePagination } from './usePagination';
+export { useMediaTrack } from './useMediaTrack';
+export { useMediaTrackByName } from './useMediaTrackByName';
