+++ packages/react/src/hooks/useMediaTrack.ts
@@ -1,0 +1,9 @@
+import { VideoSource, AudioSource } from '@livekit/components-core';
+import { UseMediaTrackOptions, useMediaTrackBySourceOrName } from './useMediaTrackBySourceOrName';
+
+export function useMediaTrack(
+  source: VideoSource | AudioSource,
+  options: UseMediaTrackOptions = {},
+) {
+  return useMediaTrackBySourceOrName({ source }, options);
+}
