+++ packages/react/src/hooks/useMediaTrackByName.ts
@@ -1,0 +1,5 @@
+import { useMediaTrackBySourceOrName, UseMediaTrackOptions } from './useMediaTrackBySourceOrName';
+
+export function useMediaTrackByName(name: string, options: UseMediaTrackOptions = {}) {
+  return useMediaTrackBySourceOrName({ name }, options);
+}
